refactor: restructure SelVeri codebase with abstract machine and high-level components
