+++ src/selveri/abstract_machine/__init__.py
@@ -1,0 +1,9 @@
+from .config import RuntimeConfiguration
+from .interpreter import ExecutionResult, interpret_ir_code, interpret_ir_text
+
+__all__ = [
+    "ExecutionResult",
+    "RuntimeConfiguration",
+    "interpret_ir_code",
+    "interpret_ir_text",
+]
