+++ src/selveri/common/__init__.py
@@ -1,0 +1,42 @@
+from .diagnostics import (
+    Counterexample,
+    Diagnostic,
+    DiagnosticCode,
+    DiagnosticFix,
+    DiagnosticLabel,
+    DiagnosticNote,
+    DiagnosticSeverity,
+    SourceFile,
+    SourceSpan,
+    TraceEntry,
+    format_found_token,
+    render_diagnostic,
+    render_expected_tokens,
+    render_internal_error,
+    should_use_color,
+)
+from .errors import SelVeriError
+from .runtime import DeclType, Scope, State, _UNSET
+
+__all__ = [
+    "Counterexample",
+    "DeclType",
+    "Diagnostic",
+    "DiagnosticCode",
+    "DiagnosticFix",
+    "DiagnosticLabel",
+    "DiagnosticNote",
+    "DiagnosticSeverity",
+    "Scope",
+    "SelVeriError",
+    "SourceFile",
+    "SourceSpan",
+    "State",
+    "TraceEntry",
+    "_UNSET",
+    "format_found_token",
+    "render_diagnostic",
+    "render_expected_tokens",
+    "render_internal_error",
+    "should_use_color",
+]
