+++ src/selveri/common/diagnostics.py
