+++ src/selveri/common/errors.py
