+++ src/selveri/compiler/__init__.py
@@ -1,0 +1,4 @@
+from .compiler import SelVeriCompiler, compile_selveri_source_to_ir_text
+from selveri.ir.instr import IRInstr
+
+__all__ = ["IRInstr", "SelVeriCompiler", "compile_selveri_source_to_ir_text"]
