+++ src/selveri/high_level/ast.py
@@ -1,0 +1,15 @@
+"""Base AST marker types for the high-level SelVeri language."""
+
+from __future__ import annotations
+
+
+class AExp:
+    pass
+
+
+class BExp:
+    pass
+
+
+class Stmt:
+    pass
