+++ src/selveri/ir/__init__.py
@@ -1,0 +1,4 @@
+from .instr import IRInstr
+from .text import parse_ir_text
+
+__all__ = ["IRInstr", "parse_ir_text"]
